+++ src/libaevol/mutation/InsertionHT.cpp
@@ -66,13 +66,13 @@ void InsertionHT::load(gzFile backup_file) {
 }
 
 void InsertionHT::generic_description_string(char* str) const {
-  sprintf(str, "%" PRId8 " %" PRId32 " %" PRId32 " %" PRId32 " %" PRId32
-      " %" PRId8 " %" PRId16 " %" PRId16 " %" PRId32 " %" PRId32,
+  sprintf(str, "%" PRId32 " %" PRId32 " %" PRId32 " %" PRId32 " %" PRId32
+      " %" PRId32 " %" PRId16 " %" PRId16 " %" PRId32 " %" PRId32,
           mut_type(),
           donor_pos1(), donor_pos2(), exogenote_pos(), receiver_pos(),
           sense(),
           donor_donor_align_.score(), exo_recv_align_.score(),
-          length_, (int32_t) -1);
+          length_, -1);
 }
 
 // ============================================================================
